pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hermes-mvp"
version = "0.4.0"
description = "Backend service for the Hermes tutoring platform"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "Hermes Team"}]
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.6",
    "sqlalchemy==2.0.34",
    "alembic==1.13.2",
    "asyncpg==0.29.0",
    "aiosqlite==0.20.0",
    "passlib[argon2]==1.7.4",
    "python-jose[cryptography]==3.3.0",
    "pydantic==2.9.2",
    "pydantic-settings==2.6.0",
    "structlog==24.4.0",
    "prometheus-fastapi-instrumentator==6.1.0",
    "opentelemetry-sdk==1.27.0",
    "opentelemetry-exporter-otlp==1.27.0",
    "opentelemetry-instrumentation-fastapi==0.48b0",
    "opentelemetry-instrumentation-sqlalchemy==0.48b0",
    "python-multipart==0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.3",
    "pytest-asyncio==0.23.8",
    "pytest-cov==5.0.0",
    "httpx==0.27.2",
    "ruff==0.6.9",
    "mypy==1.11.2",
    "types-python-jose==3.3.4.20240101",
    "types-passlib==1.7.7.20240106",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app", "app.*", "tests", "tests.*"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "B", "UP", "N", "ASYNC", "S"]
ignore = ["S101"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
strict_optional = true
show_error_codes = true
plugins = []

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=app --cov-report=term-missing"
